+++ feature/home/src/main/java/org/sopt/home/adapter/HomeWeekLinkAdapter.kt
@@ -1,8 +1,10 @@
-package org.sopt.home
+package org.sopt.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
+import org.sopt.home.viewholder.HomeWeekLinkViewHolder
+import org.sopt.home.WeekLinkDummy
 import org.sopt.home.databinding.ItemWeekLinkBinding
 import org.sopt.ui.view.ItemDiffCallback
